Select server

A helper function to conveniently switch different APIs.

Specifies the Host URL of the API to Use

kobo_host(api)

Arguments

api
Either "kobo", "kobohr", "ona", or a custom (full) URL.

Value

A single string with the URL to use.

Note

API URLs are made available for KoBo Toolbox ("kobo", https://kc.kobotoolbox.org/api/v1/), KoBo Humanitarian Response ("kobohr", https://kc.humanitarianresponse.info/api/v1/), Ona ("ona", https://ona.io/api/v1/) and Unhcr ("unhcr", https://kobocat.unhcr.org/api/v1/) . For your own installation, or other installations using the same API but accessed at a different URL, enter the full URL.

This function is not intended to be called directly. It is used in other functions.

Author

Ananda Mahto

Examples

#'
if (FALSE) {
kobo_host("unhcr")
kobo_host("ttps://kobocat.unhcr.org/api/v1/")
}