Load Packages

Load all necessary packages for koboloadeR

kobo_load_packages()

Value

No return

Examples

kobo_load_packages()

#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> ade4  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> bit64  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> car  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> caret  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> cartography  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> classInt  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> clusterSim  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> corrplot  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> data.table  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> date  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> devtools  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> DT  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> e1071  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> ellipse  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> extrafont  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> factoextra  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> FactoMineR  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> forcats  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> formatR  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> gdata  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> GGally  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> ggfortify  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> ggmap  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> ggrepel  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> ggtern  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> ggthemes  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> grid  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> gridExtra  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> gtable  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> hexbin  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> Hmisc  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> kableExtra  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> knitr  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> knncat  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> koRpus  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> lme4  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> lubridate  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> maptools  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> MASS  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> mice  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> mirt  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> moments  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> nFactors  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> pander  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> plyr  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> polycor  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> pROC  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> psych  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> qrmtools  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> quantreg  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> quantregForest  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> questionr  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> randomForest  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> raster  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> RColorBrewer  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> Rcpp  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> RCurl  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> readxl  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> REdaS  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> reshape2  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> rgdal  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> rgeos  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> rmarkdown  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> ROCR  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> rpart  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> rpart.plot  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> sampling  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> scales  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> sdcMicro  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> sp  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> stringdist  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> stringi  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> stringr  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> survey  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> tables  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> tidyverse  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> vcd  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> VIM  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> viridis  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> xgboost  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> xtable  is already installed on your computer 
#> Package Installation progress so far.. :  100%  -  100% 
#> 
#> zoo  is already installed on your computer 
#> 
#> 
#> 
#>  ***Installation completed*** 
#> 
#> 
#> 
#> 
#> 
#>  Below is the list of packages that have been now installed on your computer : 
#> 
#> ade4              Analysis of Ecological Data: Exploratory and Euclidean Metho 
#> bit64             A S3 Class for Vectors of 64bit Integers 
#> car               Companion to Applied Regression 
#> caret             Classification and Regression Training 
#> cartography       Thematic Cartography 
#> classInt          Choose Univariate Class Intervals 
#> clusterSim        Searching for Optimal Clustering Procedure for a Data Set 
#> corrplot          Visualization of a Correlation Matrix 
#> data.table        Extension of `data.frame` 
#> date              Functions for Handling Dates 
#> devtools          Tools to Make Developing R Packages Easier 
#> DT                A Wrapper of the JavaScript Library 'DataTables' 
#> e1071             Misc Functions of the Department of Statistics, Probability
#>  
#> ellipse           Functions for Drawing Ellipses and Ellipse-Like Confidence
#> R 
#> extrafont         Tools for using fonts 
#> factoextra        Extract and Visualize the Results of Multivariate Data Analy 
#> FactoMineR        Multivariate Exploratory Data Analysis and Data Mining 
#> forcats           Tools for Working with Categorical Variables (Factors) 
#> formatR           Format R Code Automatically 
#> gdata             Various R Programming Tools for Data Manipulation 
#> GGally            Extension to 'ggplot2' 
#> ggfortify         Data Visualization Tools for Statistical Analysis Results 
#> ggmap             Spatial Visualization with ggplot2 
#> ggrepel           Automatically Position Non-Overlapping Text Labels with
#> 'ggp 
#> ggtern            An Extension to 'ggplot2', for the Creation of Ternary Diagr 
#> ggthemes          Extra Themes, Scales and Geoms for 'ggplot2' 
#> grid              The Grid Graphics Package 
#> gridExtra         Miscellaneous Functions for "Grid" Graphics 
#> gtable            Arrange 'Grobs' in Tables 
#> hexbin            Hexagonal Binning Routines 
#> Hmisc             Harrell Miscellaneous 
#> kableExtra        Construct Complex Table with 'kable' and Pipe Syntax 
#> knitr             A General-Purpose Package for Dynamic Report Generation in R 
#> knncat            Nearest-neighbor Classification with Categorical Variables 
#> koRpus            An R Package for Text Analysis 
#> lme4              Linear Mixed-Effects Models using 'Eigen' and S4 
#> lubridate         Make Dealing with Dates a Little Easier 
#> maptools          Tools for Handling Spatial Objects 
#> MASS              Support Functions and Datasets for Venables and Ripley's MAS 
#> mice              Multivariate Imputation by Chained Equations 
#> mirt              Multidimensional Item Response Theory 
#> moments           Moments, cumulants, skewness, kurtosis and related tests 
#> nFactors          Parallel Analysis and Other Non Graphical Solutions to the
#> C 
#> pander            An R 'Pandoc' Writer 
#> plyr              Tools for Splitting, Applying and Combining Data 
#> polycor           Polychoric and Polyserial Correlations 
#> pROC              Display and Analyze ROC Curves 
#> psych             Procedures for Psychological, Psychometric, and Personality
#>  
#> qrmtools          Tools for Quantitative Risk Management 
#> quantreg          Quantile Regression 
#> quantregForest    Quantile Regression Forests 
#> questionr         Functions to Make Surveys Processing Easier 
#> randomForest      Breiman and Cutler's Random Forests for Classification and
#> R 
#> raster            Geographic Data Analysis and Modeling 
#> RColorBrewer      ColorBrewer Palettes 
#> Rcpp              Seamless R and C++ Integration 
#> RCurl             General Network (HTTP/FTP/...) Client Interface for R 
#> readxl            Read Excel Files 
#> REdaS             Companion Package to the Book 'R: Einführung durch angewandt 
#> reshape2          Flexibly Reshape Data: A Reboot of the Reshape Package 
#> rgdal             Bindings for the 'Geospatial' Data Abstraction Library 
#> rgeos             Interface to Geometry Engine - Open Source ('GEOS') 
#> rmarkdown         Dynamic Documents for R 
#> ROCR              Visualizing the Performance of Scoring Classifiers 
#> rpart             Recursive Partitioning and Regression Trees 
#> rpart.plot        Plot 'rpart' Models: An Enhanced Version of 'plot.rpart' 
#> sampling          Survey Sampling 
#> scales            Scale Functions for Visualization 
#> sdcMicro          Statistical Disclosure Control Methods for Anonymization of  
#> sp                Classes and Methods for Spatial Data 
#> stringdist        Approximate String Matching and String Distance Functions 
#> stringi           Character String Processing Facilities 
#> stringr           Simple, Consistent Wrappers for Common String Operations 
#> survey            Analysis of Complex Survey Samples 
#> tables            Formula-Driven Table Generation 
#> tidyverse         Easily Install and Load the 'Tidyverse' 
#> vcd               Visualizing Categorical Data 
#> VIM               Visualization and Imputation of Missing Values 
#> viridis           Default Color Maps from 'matplotlib' 
#> xgboost           Extreme Gradient Boosting 
#> xtable            Export Tables to LaTeX or HTML 
#> zoo               S3 Infrastructure for Regular and Irregular Time Series (Z's 
#> 
#> 
#> 
#> The list below indicates the missing packages: 
#> 
#> Perfect, there's no missing package 
#> 
#>