Prepare XLS form

Prepare XLSform by adding chapter, disaggregation, correlate, variable, anonymise, structuralequation, clean, cluster, predict, mappoint, mappoly in case if those fields are not exist; the function will create dummy column for each one. Also, coloring all rows that have type equal to "begin group", "end group", "begin repeat" or "end repeat".

kobo_prepare_form(form = "form.xls")

Arguments

form
The full filename of the form to be accessed (xls or xlsx file). It is assumed that the form is stored in the data folder.

Value

No return, this function edit the original XLSform directly

Examples

kobo_prepare_form()

#> 
#>  Your form should be placed within the `data` folder. 
#>  
#> Checking now for additional information within your xlsform. Note that you can insert them in the xls and re-run the function! 
#>  
#>  1- No column `labelReport` in your survey worksheet. Creating a dummy one for the moment (see readme file). ...
#> 2- Good: You have a column `variable` in your survey worksheet. This will be used to flag ordinal variable.
#> 3- Good: You have a column `disaggregation` in your survey worksheet.
#> 4- Good: You have a column `chapter` in your survey worksheet. This will be used to breakdown the generated report
#> 5- No column `structuralequation.risk` in your survey worksheet. Creating a dummy one for the moment...
#> 5- No column `structuralequation.coping` in your survey worksheet. Creating a dummy one for the moment...
#> 5- No column `structuralequation.resilience` in your survey worksheet. Creating a dummy one for the moment...
#> 6- Good: You have a column `anonymise` in your survey worksheet. This will be used to anonymise the dataset.
#> 7- Good: You have a column `correlate` in your survey worksheet. This will be used to define the variables that should be checked for correlation between each others.
#> 8- No column `clean` in your survey worksheet. Creating a dummy one for the moment (see readme file). ...
#> 9- No column `cluster` in your survey worksheet. Creating a dummy one for the moment (see readme file). ...
#> 10- No column `predict` in your survey worksheet. Creating a dummy one for the moment (see readme file). ...
#> 11- No column `mappoint` in your survey worksheet. Creating a dummy one for the moment (see readme file). ...
#> 12- No column `mappoly` in your survey worksheet. Creating a dummy one for the moment (see readme file). ...
#> 14- No column `hintReport` in your survey worksheet. Creating a dummy one for the moment (see readme file). ...
#> Workbook has no sheets!
#> 
#> ********************Survey sheet, ready to be used*********************
#>  
#> 
#>  
#>  Checking now choices sheet 
#>  
#> 1- Good: You have a column `order` in your choices worksheet.
#> 2- No column `labelReport` in your choices worksheet. Creating a dummy one for the moment...
#> 
#> ********************Choices sheet, ready to be used*********************
#>  
#> 
#>  
#>  Checking now indicator sheet 
#>  
#> 1- Good: You have a column `type` in your indicator worksheet.
#> 2- Good: You have a column `fullname` in your indicator worksheet.
#> 3- Good: You have a column `label` in your indicator worksheet.
#> 4- Good: You have a column `chapter` in your indicator worksheet.
#> 5- Good: You have a column `disaggregation` in your indicator worksheet.
#> 6- Good: You have a column `correlate` in your indicator worksheet.
#> 7- Good: You have a column `sensitive` in your indicator worksheet.
#> 8- Good: You have a column `anonymise` in your indicator worksheet.
#> 9- Good: You have a column `cluster` in your indicator worksheet.
#> 10- Good: You have a column `predict` in your indicator worksheet.
#> 11- Good: You have a column `variable` in your indicator worksheet.
#> 12- Good: You have a column `mappoint` in your indicator worksheet.
#> 13- Good: You have a column `mappoly` in your indicator worksheet.
#> 14- Good: You have a column `structuralequation` in your indicator worksheet.
#> 15- Good: You have a column `frame` in your indicator worksheet.
#> 16- Good: You have a column `listname` in your indicator worksheet.
#> 17- Good: You have a column `calculation` in your indicator worksheet.

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> 
#> ********************Indicator sheet, ready to be used*********************
#>  
#> 
#>  
#>  Checking now settings sheet 
#>  

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> 
#> ********************Settings sheet, ready to be used*********************
#>  

# NOT RUN {
kobo_prepare_form("myform.xls")
# }