Analysis project initiation

Create analysis project structure

kobo_projectinit()

Value

A structure of directory and scripts in order to set up quickly a project.

Examples

kobo_projectinit()

#> Let's create various standard folders and copy some analysis script
#> Perso exists in mainDir and is a directory.
#> Code exists in mainDir and is a directory.
#> You have multiple library path! 
#> shiny_app exists in subDir and is a directory.
#> www exists in subDir and is a directory.
#> css exists in subDir and is a directory.
#> fonts exists in subDir and is a directory.
#> image exists in subDir and is a directory.
#> Data exists in mainDir and is a directory.
#> Out directory exists in your project directory and is a directory.
#> bar_multi directory exists in out directory and is a directory.
#> disagg_multi directory exists in out directory and is a directory.
#> bar_one directory exists in out directory and is a directory.
#> disagg_one directory exists in out directory and is a directory.
#> Please open now the file called 0-config.R within the ++code++ folder and configure the path to the form and the dataset. 
#>  

#> Installing packages into ‘/home/edouard/R/x86_64-pc-linux-gnu-library/3.5’
#> (as ‘lib’ is unspecified)

#> 
#> Attaching package: ‘kableExtra’

#> The following object is masked from ‘package:dplyr’:
#> 
#>     group_rows

#> Loading required package: sylly

#> 
#> Attaching package: ‘sylly’

#> The following object is masked from ‘package:testthat’:
#> 
#>     describe

#> For information on available language packages for 'koRpus', run
#> 
#>   available.koRpus.lang()
#> 
#> and see ?install.koRpus.lang()

#> 
#> Attaching package: ‘koRpus’

#> The following object is masked from ‘package:readr’:
#> 
#>     tokenize

#> Loading required package: Hmisc

#> Loading required package: lattice

#> Loading required package: survival

#> Loading required package: Formula

#> 
#> Attaching package: ‘Hmisc’

#> The following object is masked from ‘package:koRpus’:
#> 
#>     describe

#> The following object is masked from ‘package:sylly’:
#> 
#>     describe

#> The following object is masked from ‘package:testthat’:
#> 
#>     describe

#> The following object is masked from ‘package:ape’:
#> 
#>     zoom

#> The following objects are masked from ‘package:dplyr’:
#> 
#>     src, summarize

#> The following objects are masked from ‘package:plyr’:
#> 
#>     is.discrete, summarize

#> The following objects are masked from ‘package:base’:
#> 
#>     format.pval, units

#> Loading required package: grid

#> Loading required package: Matrix

#> 
#> Attaching package: ‘Matrix’

#> The following object is masked from ‘package:tidyr’:
#> 
#>     expand

#> 
#> Attaching package: ‘survey’

#> The following object is masked from ‘package:Hmisc’:
#> 
#>     deff

#> The following object is masked from ‘package:graphics’:
#> 
#>     dotchart

#> Loading required package: viridisLite

#> Loading required package: carData

#> 
#> Attaching package: ‘car’

#> The following object is masked from ‘package:dplyr’:
#> 
#>     recode

#> Registering fonts with R

#> 
#> Attaching package: ‘zoo’

#> The following objects are masked from ‘package:base’:
#> 
#>     as.Date, as.Date.numeric

#> Google's Terms of Service: https://cloud.google.com/maps-platform/terms/.

#> Please cite ggmap if you use it! See citation("ggmap") for details.

#> 
#> Attaching package: ‘lubridate’

#> The following object is masked from ‘package:plyr’:
#> 
#>     here

#> The following objects are masked from ‘package:data.table’:
#> 
#>     hour, isoweek, mday, minute, month, quarter, second, wday, week,
#>     yday, year

#> The following object is masked from ‘package:base’:
#> 
#>     date

#> 
#> Attaching package: ‘gridExtra’

#> The following object is masked from ‘package:gdata’:
#> 
#>     combine

#> The following object is masked from ‘package:dplyr’:
#> 
#>     combine

#> 
#> Attaching package: ‘scales’

#> The following object is masked from ‘package:viridis’:
#> 
#>     viridis_pal

#> The following object is masked from ‘package:readr’:
#> 
#>     col_factor

#> corrplot 0.84 loaded