Retrieve registration data from UNHCR proGres database and generate a summary report.

A standard query that retrieve refugee registration data from proGres, aggregated at the case level. This dataset can be joined to the survey dataset in order to generate prediction model. to be used with the prediction report generation.

This includes the variables below

Arrival Date Districts of Origin Districts of Arrival Household size (case) Household size (squared) Share of members under 5 years of age Share of members between 5 and 17 years of age Share of male members between 18 and 50 Share of female members between 18 and 50 Share of members between 51 and 70 Share of members above 71 Share of members between 6 and 10 years of age Share of members between 11 and 17 years of age Share of members between 18 and 60 years of age Share of members above 60 years of age Sum of members under 5 years of age Sum of members between 6 and 10 years of age Sum of members between 11 and 17 years of age Sum of members between 18 and 60 years of age Sum of members above 60 years of age Share of members with a disability Sum of members with a disability Members above 60 years of age with a medical condition Dependency ratio Dependent members with a disability More than 3 dependents in HH

Demographics - Head of Household variables Head of HH is female Head of HH age Head above 60 years of age Head of HH is female and below 18 years of age Head of HH is disabled Head of HH education level Head of HH with a medical condition Head of HH below 18

kobo_registration()

Value

save a cleaned csv file within the data folder.

Examples

kobo_registration()

#> Error in if (nchar(uid)) st <- paste(st, ";UID=", uid, sep = ""): argument is of length zero