Generate xlsfrom skeleton from a dataframe

Creates and save a xlsform skeleton from a data.frames in your data folder The form.xls will be saved in the data folder of your project. The generated xlsfrom will need to be manually edited to configure your analysis

Note that this function only works with data.frames. The function will throw an error for any other object types.

kobo_to_xlsform(df, form = "form.xls", n = 100)

Arguments

df
dataset to use

form
The full filename of the form to be accessed (xls or xlsx file). It is assumed that the form is stored in the data folder.

n
number of levels for a factor to be considered as a text

Examples

if (FALSE) {
data(iris)
str(iris)
kobo_to_xlsform(iris)
}