Remove direct identifier

Automatically produce an anonymised dataset in line with the anonymisation plan set up in the xlsform.

This method should be used whenever Kobo or ODK forms are used as data collection tools and personal data is being collected. Even when personal data is not being collected it still may be appropriate to apply the methodology since quasi-identifiable data or other sensitive data could lead to personal identification or should not be shared. https://jangorecki.github.io/blog/2014-11-07/Data-Anonymization-in-R.html

Type			Description

----------------	-----------

Direct identifiers		Can be directly used to identify an individual. E.g. Name, Address, Date of birth, Telephone number, GPS location			Quasi- identifiers

Can be used to identify individuals when it is joined with other information. E.g. Age, Salary, Next of kin, School name, Place of work			Sensitive information

& Community identifiable information Might not identify an individual but could put an individual or group at risk. E.g. Gender, Ethnicity, Religious belief			Meta data		Data about who, where and how the data is collected is often stored separately to the main data and can be used identify individuals

The following are different anonymisation actions that can be performed on sensitive fields. The type of anonymisation should be dictated by the desired use of the data. A good approach to follow is to start from the minimum data required, and then to identify if any of those fields should be obscured.

The methods above can be reference in the column

Method			Description

----------------	--------------		Remove

Data is removed entirely from the data set. The data is preserved in the original file.		Reference		Data is removed entirely from the data set and is copied into a reference file. A random unique identifier field is added to the reference file and the data set so that they can be joined together in future. The reference file is never shared and the data is also preserved in the original file.

kobo_anonymise(frame, dico = "dico_form.xls.csv")

Arguments

dico
Generated from kobo_dico function

kobo
or odk dataset to use

Examples

# NOT RUN {
kobo_anonymise(frame, dico)
# }