Project configuration file

Write all necessary configuration files for your project

kobo_projectconfig()

Value

A file with all elements to get your data & form.

Examples

kobo_projectconfig()

#>  ##################################################################
#>  # Welcome to KoboloadeR !                                        #
#>  #an R package to facilitate data crunching of survey  microdata. #
#>  ##################################################################
#>  
#>  The 2 main advantages of KoboloadeR are to: 
#>  1. Save time to quicly generate the graphs and analysis you need to discover insights from your dataset.
#>  2. Ensure analysis reproducibilty through a separation of the analysis configuration and the analysis process.
#>  
#>  For more information: consult https://github.com/Edouard-Legoupil/koboloadeR/blob/master/README.md 
#>  or send a message to legoupil@unhcr.org
#>  
#>                      :+o. oo:. 
#>                    /ooo+  /ooo+.
#>                  /oooo/    -oooo+.
#>               .+oooo/        :osoo+- 
#>             :osoo+-             ./oooo/ 
#>           /oooo:..    -ooso+    . -+ooo+.
#>        ./ooo+. .oo/   oooooo.  -so:  /ooo+-
#>       /oooo/  .ooo:   :oooo/   .oos:  -oooo+.
#>     -ooooo/   +ooo     :oo+     +ooo   -soooo/
#>     ooooos-  :soo-   :+oooo+-    ooo+   soooos-
#>    .oooooo/.:ooo/  -oooooooooo-  .soo+.-sooooo:
#>    .oooooooooooo   /oooooooooo/   /oooosoooooo:
#>    .oooooooooos:   /oooooooooo+   .oooooooooos-
#>     ooooooooooo    /oooooooooo+    +ooooooooos-
#>     oooooooooo:    /oooooooooo+    .ooooooooos.
#>     ooooooooo+     /oooooooooo+     :oooooooos.
#>     +ooooooo+      /oooooooooo+      /oooooooo 
#>     /ooooooo.      /oooooooooo+       oooooooo 
#>     :soooooo       /oooooooooo+       /oooooo+ 
#>     .soooooo       :soooooooos/       /oooooo: 
#>      ooooooo          /oooos.         /ooooos- 
#>      +oooooo          /oooos          /ooooos  
#>      +oooooo          /oooos          /oooooo  
#>       .----.          -:::::           -----.  
#> 
#>  
#>  This script will allow you to write the configuration files for your project.
#>  
#>  
#>  First creating the necessary forlders
#>  
#>  
#>  Now configuring a few elements to write the config file: 
#>  
#>   1.- confirm the name of the main dataframe for the data. this should be a *.csv file with ; as separator 
#>  
#>   2.- Confirm the name of the xlsform file for the form. this should be a *.xls file - NOT a *.xlsx file! 
#>  
#>   3.- What sampling do you have? 

#> Error in kobo_projectconfig(): object 'runproject' not found